project/line_message_gun/requirements.txt
flask == 1.1.1
line-bot-sdk == 1.15.0
waitress >= 1.4.2
